network_data_etl/requirements.txt
pandas
httpx
pyyaml
python-dotenv
pycountry
